Moved the pull request logic out of the router into a separate file. Making both files a bit more readable
